refactor: 💡 [DSY-1686] Adjusting sample app
To Progress Indicator component
