+++ sample/src/main/kotlin/com/natura/android/sample/MainActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.natura.android.sample.components.*
 import com.natura.android.sample.patterns.ErrorActivity
-import com.natura.android.sample.patterns.LoadingActivity
+import com.natura.android.sample.components.ProgressIndicatorActivity
 import com.natura.android.sample.patterns.LogoActivity
 import com.natura.android.sample.tokens.*
 import com.natura.android.sample.tokens.icons.DrawableActivity
@@ -108,8 +108,8 @@ class MainActivity : AppCompatActivity() {
             startActivity(Intent(this, ExpandableNavigationViewActivity::class.java))
         }
 
-        btnLoader.setOnClickListener {
-            startActivity(Intent(this, LoadingActivity::class.java))
+        progressIndicatorButton.setOnClickListener {
+            startActivity(Intent(this, ProgressIndicatorActivity::class.java))
         }
 
         btnErrorDefault.setOnClickListener {
