+++ sample/src/main/kotlin/com/natura/android/sample/components/ProgressIndicatorActivity.kt
@@ -1,4 +1,4 @@
-package com.natura.android.sample.patterns
+package com.natura.android.sample.components
 
 import android.os.Bundle
 import android.view.MenuItem
@@ -6,15 +6,15 @@ import androidx.appcompat.app.AppCompatActivity
 import com.natura.android.sample.R
 import com.natura.android.sample.setChosenDefaultTheme
 
-class LoadingActivity : AppCompatActivity() {
+class ProgressIndicatorActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setChosenDefaultTheme()
 
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_loading)
+        setContentView(R.layout.activity_progress_indicator)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        supportActionBar?.title = "Loading Pattern"
+        supportActionBar?.title = "Progress Indicator"
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
