+++ sample/src/main/res/layout/activity_main.xml
@@ -200,6 +200,18 @@
             android:textColor="@color/highEmphasis" />
 
         <Button
+            android:id="@+id/progressIndicatorButton"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Progress Indicator"
+            android:drawableRight="@drawable/ds_ic_outlined_navigation_arrowleft"
+            android:textAlignment="textStart"
+            android:layout_gravity="start"
+            android:padding="16dp"
+            android:background="@color/surface"
+            android:textColor="@color/highEmphasis" />
+
+        <Button
             android:id="@+id/btnSelection"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -273,18 +285,6 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="Error Default"
-            android:drawableRight="@drawable/ds_ic_outlined_navigation_arrowleft"
-            android:textAlignment="textStart"
-            android:layout_gravity="start"
-            android:padding="16dp"
-            android:background="@color/surface"
-            android:textColor="@color/highEmphasis" />
-
-        <Button
-            android:id="@+id/btnLoader"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Loading"
             android:drawableRight="@drawable/ds_ic_outlined_navigation_arrowleft"
             android:textAlignment="textStart"
             android:layout_gravity="start"
