+++ sample/src/main/res/layout/activity_progress_indicator.xml
@@ -9,7 +9,7 @@
 
     <com.natura.android.progressindicator.ProgressIndicator
         android:id="@+id/ds_loading"
-        android:layout_width="match_parent"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:visibility="visible"
         app:layout_constraintBottom_toBottomOf="parent"
